Change principle of config loading
Now, configs are loaded when they are located in the com.mealtiger.backend.configuration.configs package and are annotated with the @Config annotation. This improves the ease of adding new configs.

Furthermore, configs are no longer accessed directly but rather through the new Configurator class. To do this the Config annotation now specifies the name of a config. Also, every method returning a config property has to be annotated with the ConfigNode annotation that specifies its property descriptor. Through these measures it is now possible to address config properties by strings such as 'Main.Database.mongoDBURL'.